feat: 모달 index로 옮김
